Split out state machines
