+++ deployer/storage.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 
 	sq "github.com/elgris/sqrl"
-	"github.com/google/uuid"
 	"github.com/pentops/o5-go/deployer/v1/deployer_pb"
 	"github.com/pentops/sqrlx.go/sqrlx"
 	"google.golang.org/protobuf/encoding/protojson"
@@ -15,12 +14,6 @@ import (
 
 var DeploymentNotFoundError = fmt.Errorf("deployment not found")
 var StackNotFoundError = fmt.Errorf("stack not found")
-
-var namespaceStackID = uuid.MustParse("C27983FD-BC4B-493F-A056-CC8C869A1999")
-
-func StackID(envName, appName string) string {
-	return uuid.NewMD5(namespaceStackID, []byte(fmt.Sprintf("%s-%s", envName, appName))).String()
-}
 
 func getDeployment(ctx context.Context, tx sqrlx.Transaction, id string) (*deployer_pb.DeploymentState, error) {
 	var deploymentJSON []byte
