+++ service/service.go
@@ -8,8 +8,8 @@ import (
 
 	sq "github.com/elgris/sqrl"
 	"github.com/pentops/log.go/log"
-	"github.com/pentops/o5-deploy-aws/deployer"
 	"github.com/pentops/o5-deploy-aws/github"
+	"github.com/pentops/o5-deploy-aws/states"
 	"github.com/pentops/o5-go/deployer/v1/deployer_spb"
 	"github.com/pentops/o5-go/deployer/v1/deployer_tpb"
 	"github.com/pentops/outbox.pg.go/outbox"
@@ -63,7 +63,7 @@ func NewDeployerService(conn sqrlx.Connection, github github.IClient) (*Deployer
 	}
 
 	deploymentQuery, err := deployer_spb.NewDeploymentPSMQuerySet(
-		deployer_spb.DefaultDeploymentPSMQuerySpec(deployer.DeploymentTableSpec().StateTableSpec()),
+		deployer_spb.DefaultDeploymentPSMQuerySpec(states.DeploymentTableSpec().StateTableSpec()),
 		psm.StateQueryOptions{},
 	)
 	if err != nil {
@@ -71,7 +71,7 @@ func NewDeployerService(conn sqrlx.Connection, github github.IClient) (*Deployer
 	}
 
 	stackQuery, err := deployer_spb.NewStackPSMQuerySet(
-		deployer_spb.DefaultStackPSMQuerySpec(deployer.StackTableSpec().StateTableSpec()),
+		deployer_spb.DefaultStackPSMQuerySpec(states.StackTableSpec().StateTableSpec()),
 		psm.StateQueryOptions{},
 	)
 	if err != nil {
