+++ states/deployment.go
@@ -1,4 +1,4 @@
-package deployer
+package states
 
 import (
 	"context"
